+++ ports/jetlancer/jetlancer/tools/patchscript
@@ -1,0 +1,160 @@
+#!/bin/bash
+# Set GAMEDIR to the current directory and set logfile
+GAMEDIR="$PWD"
+LOGFILE="$GAMEDIR/patchlog.txt"
+
+# Redirect output and error to the log file
+exec > >(tee -a "$LOGFILE") 2>&1
+echo "GAMEDIR is set to: $GAMEDIR"
+
+# Exports
+export DATADIR="$GAMEDIR/gamedata"
+export LD_LIBRARY_PATH="/usr/lib32:$GAMEDIR/libs:$GAMEDIR/tools/libs:$LD_LIBRARY_PATH"
+export SDL_GAMECONTROLLERCONFIG="$sdl_controllerconfig"
+export TOOLDIR="$GAMEDIR/tools"
+export TMPDIR="$GAMEDIR/tmp"
+export PATH="$GAMEDIR/tools:$PATH"
+
+# Permissions
+chmod 666 /dev/uinput
+chmod 777 "$TOOLDIR/gmKtool.py"
+chmod 777 "$TOOLDIR/oggenc"
+chmod 777 "$TOOLDIR/oggdec"
+
+cd "$GAMEDIR"
+
+apply_xdelta() {
+# If "gamedata/data.win" exists and matches the checksum of the GOG or Steam versions
+    if [ -f "./gamedata/data.win" ]; then
+        checksum=$(md5sum "./gamedata/data.win" | awk '{print $1}')
+    
+        # Checksum for the GOG version
+        if [ "$checksum" = "0258cadce342712c2ffbd5fc70e0dfd7" ]; then
+        $ESUDO $TOOLDIR/xdelta3 -d -s gamedata/data.win -f ./tools/patches/gogdata.xdelta gamedata/game.droid && \
+        rm gamedata/data.win
+	echo "GOG data.win has been patched"
+        # Checksum for the Steam version
+        elif [ "$checksum" = "86045b6464bb909851634c3e34b2a82e" ]; then
+            $ESUDO $TOOLDIR/xdelta3 -d -s gamedata/data.win -f ./tools/patches/steamdata.xdelta gamedata/game.droid && \
+            rm gamedata/data.win
+ 	    echo "Steam data.win has been patched"
+        else
+            echo "Error: MD5 checksum of data.win does not match any expected version."
+	    exit 1
+        fi
+    else    
+        echo "Error: Missing data.win in gamedata folder or game has been patched."
+    fi
+
+    # If "gamedata/audiogroup1.dat" exists and matches the checksum of the Steam version
+    if [ -f "./gamedata/data.win" ]; then
+        checksum=$(md5sum "./gamedata/audiogroup1.dat" | awk '{print $1}')
+    
+        # Checksum for the Steam audiogroup1.dat version
+        if [ "$checksum" = "bcadc44e45f8a6caf9b5c5db538ced13" ]; then
+        $ESUDO $TOOLDIR/xdelta3 -d -s gamedata/audiogroup1.dat -f ./tools/patches/steamaudiogroup1.xdelta gamedata/audiogroup1patched.dat && \
+        rm gamedata/audiogroup1.dat
+        mv gamedata/audiogroup1patched.dat gamedata/audiogroup1.dat
+        else
+            echo "Error: MD5 checksum of audiogroup1.dat does not match any expected version."
+        fi
+    else    
+        echo "Error: Missing audiogroup1.dat in gamedata folder or game has been patched."
+    fi
+}
+
+compress_audio() {
+    # Compress audio
+    if [ -f "$GAMEDIR/compress.txt" ]; then
+        echo "Compressing audio. The process will take 5-10 minutes"  > $CUR_TTY
+        mkdir -p "$TMPDIR"
+        ./tools/gmKtool.py -v -m 262144 -b 64 -d "$TMPDIR" "$GAMEDIR/gamedata/game.droid"
+
+        if [ $? -eq 0 ]; then
+            mv $TMPDIR/* "$GAMEDIR/gamedata"
+            rm "$GAMEDIR/compress.txt"
+            rmdir "$TMPDIR"
+            echo "Audio compression applied successfully." > $CUR_TTY
+        else
+            echo "Audio compression failed." > $CUR_TTY
+            rm -rf "$TMPDIR"
+        fi
+    fi
+
+    sleep 3
+
+    # Find and compress all .ogg files in the /gamedata directory
+    find "$GAMEDIR/gamedata" -type f -name "*.ogg" | while read -r file; do
+    # Create a temporary file for the compressed version
+    temp_file="${file%.ogg}_temp.ogg"
+
+    # Decode the .ogg file and encode it with the specified bitrate
+    "$TOOLDIR/oggdec" -o - "$file" | "$TOOLDIR/oggenc" -b 64 -o "$temp_file" -
+    if [ $? -eq 0 ]; then
+        # Replace the original file with the compressed version if successful
+        mv "$temp_file" "$file"
+        echo "Compressed: $file to 64 kbps"
+    else
+        echo "Failed to compress: $file" >&2
+        rm -f "$temp_file" # Clean up the temp file in case of error
+    fi
+    done
+
+    echo "All .ogg files have been processed."
+}
+
+process_game() {
+       local files_to_delete=(
+      "./gamedata/Galaxy.dll"
+      "./gamedata/GOG.gml.dll"
+      "./gamedata/gog.ico"
+      "./gamedata/goggame-1489231569.hashdb"
+      "./gamedata/goggame-1489231569.ico"
+      "./gamedata/goggame-1489231569.info"
+      "./gamedata/goggame-galaxyFileList.ini"
+      "./gamedata/goglog.ini"
+      "./gamedata/Jet Lancer.exe"
+      "./gamedata/Launch Jet Lancer.lnk"
+      "./gamedata/support.ico"
+      "./gamedata/unins000.dat"
+      "./gamedata/unins000.exe"
+      "./gamedata/unins000.ini"
+      "./gamedata/unins000.msg"
+      "./gamedata/steam_api.dll"
+    )
+
+    # Delete unnecessary files
+    for file in "${files_to_delete[@]}"; do
+      if [ -f "$file" ]; then
+        rm "$file"
+        echo "Deleted $file"
+      else
+        echo "$file not found"
+      fi
+    done
+
+    apply_xdelta
+    compress_audio
+    sleep 3
+    
+    # Check for .ogg files and move to APK
+    if [ -n "$(ls ./gamedata/*.ogg 2>/dev/null)" ]; then
+        mkdir -p ./assets
+        mv ./gamedata/*.ogg ./assets/
+        echo "Moved .ogg files to ./assets/"
+
+        zip -r -0 ./jetlancer.apk ./assets/
+        echo "Zipped contents to ./jetlancer.apk"
+
+        rm -rf ./assets
+        echo "Deleted assets directory"
+    else
+        echo "No .ogg files found"
+    fi
+	
+    # Final completion message
+    echo "Patching process complete!"
+}
+
+# Call the function
+process_game
